requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
